src/api/requirements.txt
fastapi
pydantic
python-dotenv
azure-ai-projects
azure-identity
uvicorn
gunicorn
pandas
pyodbc
aiohttp
azure-ai-vision-imageanalysis
semantic-kernel
semantic-kernel[azure]